controlles are divided + some fixes
