+++ admin/src/app/Components/hotels-filter/hotels-filter.component.html
@@ -12,7 +12,7 @@
                 *ngFor="let name of names" 
                 [value]="name"
             >
-                {{name}}
+                {{ name }}
             </mat-option>
         </mat-autocomplete>
     </mat-form-field>
